pixelwands are not memory managed
The problem with putting a finalizer on a pixelwand is, I think, that
pixelwands are often passed into MagickWand where anything may happen with
them.  Chicken may try to garbage collect them when they are still needed.
It would be nice to come up with a better solution to this problem.