image datatype no longer exposed via api
The Image datatype is not safe to expose to scheme without careful
consideration of its memory management.  Its only use in MagickWand is
the construction of new magickwands from a given list of
images (references to which have been obtained from other magickwands).

Memory management of images is done via internal reference counting, but
simply obtaining a reference via GetImageFromMagickWand does not
increase the refcount, so it would be possible for an image to be
deallocated internally within MagickWand while a reference still exists
on the scheme side.

If we want to be able to use the Image datatype in scheme, we will need
a wrapper that increments the image's refcount whenever a reference is
taken to the scheme side.  This, we leave as an exercise for another
day.